requirements_uv.txt
# Requirements file for UV environment compatibility
# This file contains all dependencies from uv_packages.txt
# For UV users, use pyproject.toml instead

absl-py
accelerate
aiohappyeyeballs
aiohttp
aiohttp-cors
aiosignal
airportsdata
annotated-types
antlr4-python3-runtime
anyio
astor
async-timeout
attrs
blake3
cachetools
certifi
charset-normalizer
click
cloudpickle
colorful
compressed-tensors
contourpy
cycler
datasets
deepspeed
depyf
dill
diskcache
distlib
distro
docker-pycreds
einops
fastapi
filelock
flash_attn
fonttools
frozenlist
fsspec
gguf
gitdb
GitPython
google-api-core
google-auth
googleapis-common-protos
grpcio
h11
hf_transfer
hjson
htcondor
httpcore
httptools
httpx
huggingface-hub
idna
iniconfig
interegular
Jinja2
jiter
jsonschema
jsonschema-specifications
kiwisolver
lark
latex2sympy2_extended
lm-format-enforcer
Markdown
markdown-it-py
MarkupSafe
math-verify
matplotlib
mdurl
mistral_common
mpmath
msgpack
msgspec
multidict
multiprocess
networkx
ninja
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-ml-py
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
openai
opencensus
opencensus-context
opencv-python-headless
outlines
outlines_core
pandas
partial-json-parser
peft
pillow
pluggy
prometheus-fastapi-instrumentator
prometheus_client
propcache
proto-plus
protobuf
psutil
py-cpuinfo
py-spy
pyarrow
pyasn1
pyasn1_modules
pybind11
pycountry
pydantic
pydantic_core
pyparsing
pytest
python-dateutil
python-dotenv
pytz
PyYAML
ray
referencing
regex
requests
rich
rpds-py
rsa
safetensors
sentencepiece
sentry-sdk
setproctitle
smart-open
smmap
sniffio
starlette
sympy
tensorboard
tensorboard-data-server
tiktoken
tokenizers
tomli
torch
torchvision
tqdm
transformers
triton
trl
typing_extensions
tzdata
urllib3
uvicorn
uvloop
virtualenv
vllm
wandb
watchfiles
websockets
Werkzeug
wrapt
xformers
xgrammar
xxhash
yarl 